+++ OSLoader/HAL/board_up.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #include "FreeRTOS.h"
 
-uint32_t nsToCycles(uint64_t nstime, uint64_t period, uint64_t min) ;
+uint64_t nsToCycles(uint64_t nstime, uint64_t period, uint64_t min) ;
 
 bool driverWaitTrueF(bool (*f)(), TickType_t timeout);
 bool driverWaitFalseF(bool (*f)(), TickType_t timeout);
